botv1/requirements.txt

discord.py==2.5.2
asyncpg==0.30.0
flask==3.0.0
requests==2.31.0
PyNaCl>=1.5.0
